move unused pod validator code to skaffold
